chore: adjust recitation_tracks bulk upload to coding standard
