--- apps/content/api/portal/schemas/__init__.py
@@ -1,4 +1,0 @@
-from apps.core.ninja_utils.router import ItqanRouter
-
-# Kept to satisfy broad portal router autodiscovery.
-router = ItqanRouter()
